Change IP lookup method
I used 2 ip lookup services and it seems like they return an empty string on a github runner.
Therefore I am trying a switch to `dig` to lookup the IP of the runner to be allowed through the firewall.
There's a chance the runner replaces any variable that contains sensitive info with empty strings. But this is worth a shot.